+++ environment.yml
@@ -7,7 +7,6 @@ dependencies:
 - numpy
 - pandas
 - alembic
-- paramiko
 - pip:
   - versioneer
   - twine
