+++ examples/experiment/environment.yml
@@ -1,6 +1,7 @@
 name: provenance-experiment-example
 dependencies:
   - python==3.5
+  - paramiko
   - pip:
       - "git+https://github.com/bmabey/provenance.git"
       - pyyaml
