+++ setup.py
@@ -9,10 +9,11 @@ setup(
     name='provenance',
     version=versioneer.get_version(),
     cmdclass=versioneer.get_cmdclass(),
-    packages=('provenance',),
+    packages=['provenance', 'provenance.sftp'],
     setup_requires=['pytest>=3.0.0', 'pytest-runner'],
     install_requires=[open('requirements.txt').read().strip().split('\n')],
     tests_requires=[open('test_requirements.txt').read().strip().split('\n')],
+    extras_require={'sftp': ['paramiko']},
     include_package_data=True,
     description="Provenance and caching library for functions",
     long_description=(open('README.rst').read() if exists('README.rst')
